Move initial integration test to phpunit and execute integration tests in travis
