+++ .travis.yml
@@ -18,9 +18,10 @@ php:
 
 script:
     - composer install
-    - ./vendor/bin/phpunit src/
+    - ./vendor/bin/phpunit src/OpenCATS/Tests/UnitTests
     - cd docker/; docker-compose -f docker-compose-test.yml up -d
     - docker-compose -f docker-compose-test.yml run -w /var/www/public php dockerize -wait tcp://mariadb:3306 -wait http://opencats:80 -timeout 30s
     - docker-compose -f docker-compose-test.yml run -w /var/www/public php php modules/tests/waitForDb.php
     - docker-compose -f docker-compose-test.yml run -w /var/www/public php cat config.php
+    - docker-compose -f docker-compose-test.yml run -w /var/www/public/ php ./vendor/bin/phpunit src/OpenCATS/Tests/IntegrationTests
     - docker-compose -f docker-compose-test.yml run -w /var/www/public php ./vendor/bin/behat -v
